Use scipy's built-in next-optimal-fft-length function instead of my round2()
Depending on the BLAS you have there are myriad optimal sizes. [For example](https://github.com/scipy/scipy/pull/3144#issuecomment-68151925)
the optimal length under one implementation is any of 2^a 3^b 5^c 7^d 11^e 13^f,
where e+f is either 0 or 1, and the other exponents are arbitrary. It's just a
matter of how the implementation chooses to carve up subproblems. Halves---making
powers of 2 fast---is traditional, but not the only choice.