Use pip install soundfile in GitHub Actions
Maybe this is why codecov shows identical tests with and without this?

`pip install pysoundfile` installs an older version that doesn't support
mp3, so it should be failing tests and isn't.

[Ah, the old version of pysoundfile is installing, but failing to be
imported, so then it falls back to scipy, whether soundfile is installed
or not, which explains why the coverage files were identical for both
cases!]